simulationGraph htmlNode 함수 분리
