+++ src/constants/simulationGraphNodePositions.js
@@ -1,0 +1,36 @@
+export const SIMULATION_GRAPH_NODE_POSITIONS = {
+  매출원가: { x: 798.3342288998895, y: 704.8898453135909 },
+  매출액: { x: 798.3342288998895, y: 968.24288111116 },
+  당기제품제조원가: { x: 1347.24006, y: 347.0353200556658 },
+  기초재고: { x: 1374.8861906279813, y: 525.3371253720223 },
+  기말재고: { x: 1401.164721046636, y: 705.6113923763377 },
+  FERT100s: { x: 1404.4980606279814, y: 880.7656582299755 },
+  FERT200s: { x: 1402.8313908373088, y: 1058.8715207419013 },
+  당기제조비용: { x: 1948.7291706279816, y: 64.39271855322282 },
+  재공품: { x: 1987.308, y: 247.7019956226425 },
+  액티비티단수차: { x: 1987.3079999999993, y: 435.6673726920622 },
+  액티비티배부: { x: 1948.7291706279816, y: 625.0334101801366 },
+  가공비: { x: 2456.55923882003, y: -401.6180286126659 },
+  FERT201: { x: 2011.91221064616, y: 1196.5029406541853 },
+  FERT202: { x: 1986.6887174199474, y: 1427.2064964627455 },
+  FERT203: { x: 1936.2417309675197, y: 1667.0822316262913 },
+  부재료비: { x: 2487.364425998405, y: -216.56017066438633 },
+  원재료비: { x: 2546.731658653151, y: -34.56237434227623 },
+  생산입고: { x: 2463.3286336951332, y: 325.4501012670632 },
+  공정출고: { x: 2466.6619670284667, y: 170.44829371390944 },
+  액티비티수차합: { x: 2502.310186322719, y: 446.57359373876614 },
+  액티비티단가합: { x: 2449.995481117116, y: 625.4181075762423 },
+  비용계획합: { x: 2903.202914105695, y: -401.01035949159933 },
+  FERT101: { x: 2946.6635023857293, y: 888.9510872451323 },
+  FERT102: { x: 2870.3583398903847, y: 1115.7534849313272 },
+  FERT103: { x: 2775.78644077191, y: 1352.4690521094726 },
+  FERT104: { x: 2689.881831225508, y: 1581.7646718201397 },
+  FERT105: { x: 2604.9951251342463, y: 1806.937392717706 },
+  FERT106: { x: 2540.464402190489, y: 2015.3884534790848 },
+  ROH0001누적: { x: 3473.87079917938, y: 2.6317058812969503 },
+  ROH0002누적: { x: 3439.8521835017586, y: 214.00036899174603 },
+  ROH0003누적: { x: 3386.574548317458, y: 435.4222352384621 },
+  ROH2001누적: { x: 3429.372069069061, y: -617.4707944957579 },
+  ROH2002누적: { x: 3470.3554648643294, y: -416.9836287512874 },
+  ROH2003누적: { x: 3526.152592648285, y: -218.4055762519053 },
+};
